bugfix: create missing role that led to test failure
The test checks for a "ProvidedRole"-connection between "BasicComponent
BC1" and "Interface IP1". This role was not previously created by the
test because of missing coordinates and wrong control-dimensions as
params for the mouse actions. This is now fixed.

I tried to keep the commands as simple as possible, but it seems to be
not possible to completly do without mouse-coordinates or at least
control-dimensions.